requirements.txt
bs4
trafilatura
lyricsgenius

pandas
djangorestframework